// types/project.ts
export interface Project {
  id: string
  projectTitle: string
  projectNumber: string
  projectDescription: string
  projectType: string
  location: string
  estimatedBudget: number
  startDate: string
  endDate: string
  departmentResponsible: string
  status: string
  createdAt: string
  updatedAt: string
}